+++ src/iminuit/cost.py
@@ -61,6 +61,9 @@ def _sum_z_squared_soft_l1(y, ye, ym):
     return np.sum(2 * (np.sqrt(1.0 + z) - 1.0))
 
 
+# If numba is available, use it to accelerate computations in float32 and float64
+# precision. Fall back to plain numpy for float128 which is not currently supported
+# by numba.
 try:
     import numba as nb
     from numba.extending import overload
@@ -69,19 +72,19 @@ try:
 
     @overload(_spd_transform, inline="always")
     def _spd_transform_ol(n, mu):
-        return _spd_transform_np  # pragma: nocover
+        return _spd_transform_np  # pragma: no cover
 
     _log_poisson_part_np = _log_poisson_part
 
     @overload(_log_poisson_part, inline="always")
     def _log_poisson_part_ol(n, mu):
-        return _log_poisson_part_np  # pragma: nocove
+        return _log_poisson_part_np  # pragma: no cover
 
     _z_squared_np = _z_squared
 
     @overload(_z_squared, inline="always")
     def _z_squared_ol(y, ye, ym):
-        return _z_squared_np  # pragma: nocove
+        return _z_squared_np  # pragma: no cover
 
     _sum_log_x_np = _sum_log_x
     _sum_log_x_nb = nb.njit(
@@ -147,6 +150,7 @@ try:
         return _sum_z_squared_soft_l1_np(y, ye, ym)
 
     del nb
+    del overload
 except ModuleNotFoundError:  # pragma: no cover
     pass